ChatProject/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.2.0
pytz==2023.3.post1
service-identity==23.1.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
zope.interface==6.1